refactor(polytope_playground): extract cross-section/surface policy into sections module
Moves SurfaceMode, SectionLayer, section_layer_projection, and the heavy-polychora SDF gate (row_blocks_sdf) out of state.rs into sections.rs, re-exported from state like projections. Also demotes cross-module Demo references in projections.rs/sections.rs to plain code spans (they target Demo fields/private methods across the private state module, which intra-doc links can't resolve). state.rs 2098 -> ~1490; clippy -D clean, 97 tests, 0 doc warnings.
